Feature/vb006 - Likes (Bucket approach) (#6)
* Swapped to FieldValue for updating arrays

* Updated interfaces for FieldValue

* expensive like checking

* auto refresh list on likes update

* firebase function for aggregating recent likes

* Listener on likes subcollection

* removed aggregation cloud func for now

* likes saved onto idea doc as array

* first draft of Likes Bucket

* likes bucket

+++ src/components/Idea.vue
@@ -1,15 +1,18 @@
 <template>
 	<div class="mb-5">
-		<div class="list bg-gray-300 rounded text-blue-800 p-5">
+		<div v-if="!loading && idea.data" class="list bg-gray-300 rounded text-blue-800 p-5">
 			<li>description: {{ idea.data.description }}</li>
 			<li>benefit: {{ idea.data.benefit }}</li>
 			<li>skills required: {{ readableSkills }}</li>
 			<li>opened: {{ idea.data.createdOn.toDate() }}</li>
-			<li>likes: {{ idea.data.likes.length }}</li>
+			<li>likes: {{ idea.data.likesCount }}</li>
 			<li>status: {{ readableStatus }}</li>
 			<br />
-			<li>Owner: {{ ownerName }}</li>
+			<li>Owner: {{ this.idea.data.owner.name }}</li>
+			<button v-if="!userHasLiked" class="like" @click="idea.like()">Like</button>
+			<button v-else class="like" @click="idea.unlike()">Unlike</button>
 			<button class="close" @click="deleteIdea()">Delete</button>
+			{{ recentLikes }}
 		</div>
 	</div>
 </template>
@@ -17,42 +20,55 @@
 <script lang="ts">
 import { Component, Prop, Vue } from "vue-property-decorator";
 import { User, Idea } from "@/models";
-import { Skill, IdeaStatus } from "@/models/typings";
+import { Skill, IdeaStatus, Like } from "@/models/typings";
+import { State } from "vuex-class";
+import { db } from "../firebase";
+import { Likes } from "../models";
 
 @Component
 export default class IdeaComponent extends Vue {
+	// Mapped Store
+	@State user: User;
+
 	// Props
 	@Prop() idea: Idea;
 
 	// Data
-	owner: User = null;
+	loading: boolean = true;
+	likes: Likes = null;
 
 	// Hooks
-	async created() {
-		this.owner = await new User(this.idea.data.owner).init();
-	}
+	async mounted() {
+		this.likes = await new Likes(this.idea.id).init();
 
-	destroyed() {
-		this.owner.unsubscribe();
+		this.loading = false;
 	}
 
 	// Methods
 	deleteIdea() {
-		this.idea.delete();
-		this.$destroy();
+		if (confirm("Are you sure?")) {
+			this.idea.delete();
+		}
 	}
 
 	// Computed
 	get readableSkills() {
-		return this.idea.data.skillsRequired.map(skill => Skill[skill]).toString();
+		return this.idea.data.skillsRequired.map(skill => Skill[skill]).join(", ");
 	}
 
 	get readableStatus() {
 		return IdeaStatus[this.idea.data.status];
 	}
 
-	get ownerName() {
-		return this.owner && this.owner.data ? this.owner.data.name : "Loading...";
+	get recentLikes() {
+		return this.likes.latest
+			.splice(0, 2)
+			.map(like => (like.ref.id === this.user.id ? "You" : like.name))
+			.join(", ");
+	}
+
+	get userHasLiked() {
+		return this.likes.ids.some(id => id === this.user.id);
 	}
 }
 </script>
@@ -62,7 +78,17 @@ export default class IdeaComponent extends Vue {
 	@apply capitalize;
 }
 
+.button {
+	@apply text-lg text-white font-semibold px-6 py-1 rounded mt-2;
+}
+
+.like {
+	@apply button;
+	@apply bg-green-500;
+}
+
 .close {
-	@apply bg-red-400 text-white text-lg font-semibold px-6 py-1 rounded mt-2;
+	@apply button;
+	@apply bg-red-400 float-right;
 }
 </style>
